\name{avaliaR-package}
\alias{avaliaR-package}
\alias{avaliaR}
\docType{package}
\title{
\packageTitle{avaliaR}
}
\description{
\packageDescription{avaliaR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{avaliaR}
\packageIndices{avaliaR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{avaliaR}

Maintainer: \packageMaintainer{avaliaR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
