% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escore.R
\name{escore}
\alias{escore}
\title{Escore}
\usage{
escore(respostas, caderno, estrutura_prova, nao_apres = NULL)
}
\arguments{
\item{respostas}{vetor de caracteres com as respostas dos alunos em cada um 
dos itens do caderno de provas.}

\item{caderno}{vetor com o número do caderno que o participante respondeu.}

\item{estrutura_prova}{objeto resultante da função \code{estrutura_prova}.}

\item{nao_apres}{caracter de tamanho 1 contendo o código para os itens não 
apresentados, ex.: \code{'9'}}
}
\value{
ATENÇÃO: O resultado da função está ordenado por item.
A saída da função é um \code{data.frame} com 7 colunas:
\describe{
  \item{respostas}{string com as resposta dos alunos em cada um dos itens da prova (todos os cadernos),
  respeitando o sequencial do item em \code{estrutura_prova}.}
  \item{padrao}{padrão de respostas considerando a codificação definida no \code{estrutura_prova}.}
  \item{ntot}{Número total de itens (considerando todos os possíveis cadernos).}
  \item{nnao_apres}{Número de itens não apresentados ao participante.}
  \item{ntried}{Quantidade de itens que foi respondida pelo partipante.}
  \item{score}{Soma da pontuação do participante em cada item.}
  \item{pacer}{percentual de acertos}
}
}
\description{
Esta função tem por objetivo calcular o escore dos participantes.
}
\seealso{
estrutura_prova
}
