% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrutura_prova.R
\name{estrutura_prova}
\alias{estrutura_prova}
\title{Estrutura da Prova}
\usage{
estrutura_prova(itens_caderno, coditem, respostas, tipo_item, pontuacao,
  gabarito, nit)
}
\arguments{
\item{itens_caderno}{objeto da classe 'itens_caderno', gerado a partir da funcao 
\code{gera_itens_caderno}.}

\item{coditem}{vetor com códigos dos itens.}

\item{respostas}{objeto do tipo \code{list} onde cada elemento é o conjunto
de opções de respostas possíveis. \code{Ex.: c(LETTER[1:4], '*', 'c')}.}

\item{tipo_item}{objeto do tipo \code{vector} indicando para cada item se ele é 
\code{'dicotomico'} ou \code{'politomico'}.}

\item{pontuacao}{objeto do tipo \code{list} indicando a pontuacao para os casos de 
itens politômicos. Em caso de item dicotômico declarar NULL.
Exemplo: \code{list(NULL, c(0.0, 0.5, 1.0))}}

\item{gabarito}{objeto do tipo \code{vetor} indicando os gabaritos para os itens 
dicotômicos. Em caso de item politômico declarar NULL. Exemplo: \code{c("A", NULL)}}

\item{nit}{Número de itens que estão sendo adicionados ou criados com essa função.
o nit deve ser igual ao número de elementos em cada um dos parâmetros anteriores. 
Observação: caso só existam itens dicotômicos então \code{pontuacao} 
pode não ser declarada. Default: \code{NULL}
Observação: caso só existam itens politômicos então \code{gabarito} 
pode não ser declarada. Default: \code{NULL}}

\item{nit}{numero de itens total}
}
\description{
Esta função tem por objetivo gerar um objeto da classe 'estrutura_teste'
   \code{gera_gabarito} e \code{ItemPos}.
}
\seealso{
gera_itens_caderno
}
