% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_gabarito.R
\name{gera_gabarito}
\alias{gera_gabarito}
\title{Funcao que gera lista de gabaritos e pontuacao para o calculo do escore e estatisticas classicas.}
\usage{
gera_gabarito(coditem, respostas, tipo_item, pontuacao = NULL,
  gabarito = NULL, nit, data = NULL)
}
\description{
Esta funcao tem por objetivo gerar um objeto do tipo list que sera utilizado
   dentro da funcao escore e tambem da funcao de tct. 
 
 @param coditem vetor com códigos dos itens.
 @param respostas objeto do tipo \code{list} onde cada elemento é o conjunto
    de opções de respostas possíveis. \code{Ex.: c(LETTER[1:4], '*', 'c')}.
 @param tipo_item objeto do tipo \code{vector} indicando para cada item se ele é 
    \code{'dicotomico'} ou \code{'politomico'}.
 @param pontuacao objeto do tipo \code{list} indicando a pontuacao para os casos de 
    itens politômicos. Em caso de item dicotômico declarar NULL.
    Exemplo: \code{list(NULL, c(0.0, 0.5, 1.0))}
 @param gabarito objeto do tipo \code{vetor} indicando os gabaritos para os itens 
    dicotômicos. Em caso de item politômico declarar NULL. Exemplo: \code{c("A", NULL)}
 @param nit Número de itens que estão sendo adicionados ou criados com essa função.
    o nit deve ser igual ao número de elementos em cada um dos parâmetros anteriores. 
    Observação: caso só existam itens dicotômicos então \code{pontuacao} 
    pode não ser declarada. Default: \code{NULL}
    Observação: caso só existam itens politômicos então \code{gabarito} 
    pode não ser declarada. Default: \code{NULL}
 @param data parâmetro opcional. Default: \code{NULL}. Caso esteja adicionando 
    novos itens à um gabarito já existente então o gabarito existente pode
    ser definido em \code{data = gabarito_existente}. Assim, os itens serão 
    acrescentados após o \code{gabarito_existente}.
}
\examples{
gera_gabarito(coditem = paste0("CODIGO",sprintf("\%05d", 1:5)),
              respostas = c(list(c(LETTERS[1:3])), rep(list(c(LETTERS[1:4], '*', '.')), 3), list(c(LETTERS[3:1]))),
              tipo_item = c('politomico',rep('dicotomico', 3), 'politomico'), 
              pontuacao = list(c(0.0,0.5,1.0), c(0, 1), c(0, 0.5), NULL, c(0,1,2)),
              gabarito = c(NA,"A","C","B",NA),
              nit = 5,
              data = NULL)
}
