% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_itens_caderno.R
\name{gera_itens_caderno}
\alias{gera_itens_caderno}
\title{Construtor de Itens por Caderno}
\usage{
gera_itens_caderno(...)
}
\description{
Funcao para construir o objeto da classe \code{'itens_caderno'} que 
  sera utilizado na geracao da estrutura da prova \code{estrutura_prova}.
 
 @param ... vetores com o sequencial unico do item. Assume-se que os 
 cadernos estão sendo dispostos na ordem em que serao utilizados. 
 A ordenacao dos itens deve ser a mesma utilizada nos cadernos
 de prova fisicos.
 @seealso estrutura_prova
}
