% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItemPos.R
\name{ItemPos}
\alias{ItemPos}
\title{Cria lista para informar a posicao do item nos cadernos}
\usage{
ItemPos(itens, nitens, tipo)
}
\arguments{
\item{itens}{Objeto do tipo matriz com \code{rownames} e \code{colnames} definido.}

\item{nitens}{Numero de itens total.}

\item{tipo}{determina o tipo de output \code{integer} ou \code{character}}
}
\value{
Objeto do tipo \code{list} de tamanho \code{nitens} onde cada elemento eh um item.
  Cada objeto da lista eh um objeto do tipo \code{numeric matrix}.
  Cada linha desta matriz refere-se ao caderno em que o
  item aparece e contem a informacao da posicao deste item no caderno.
}
\description{
Cria lista para informar a posicao do item nos cadernos
}
\examples{
itens_teste <- matrix(c(1:30,11:30,1:10,21:30,1:20),ncol=30,byrow = TRUE)
rownames(itens_teste) <- 1:nrow(itens_teste)
colnames(itens_teste) <- paste('it',sprintf('\%02d',1:30),sep='')
#saida numerica
ItemPos(itens = itens_teste, nitens = 30,tipo='integer')

#saida como character
ItemPos(itens = itens_teste, nitens = 30,tipo='character')

}
