% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MostraPaleta.R
\name{MostraPaleta}
\alias{MostraPaleta}
\title{Exibe Paleta de cores}
\usage{
MostraPaleta(cores)
}
\arguments{
\item{cores}{Vetor com as cores que compoem a paleta de cores escolhida}
}
\description{
Exibe Paleta de cores
}
\examples{

## 
 MostraPaleta(c('blue','cyan','red','yellow'))
 
}
