% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posicao_itens.R
\name{posicao_itens}
\alias{posicao_itens}
\title{Posicao dos itens}
\usage{
posicao_itens(itens_caderno, nit)
}
\description{
Funcao para construir uma lista onde cada elemento eh um item na ordem 
  sequencial e indica a posicao do item nos diferentes cadernos de prova.
 
 @param itens_caderno objeto da classe \code{itens_caderno}.
 @param nit numero de itens total.
 @seealso estrutura_prova, gera_itens_caderno
 @export
}
