% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_itemgab_print.R
\name{print.info_itemgab}
\alias{print.info_itemgab}
\title{Print Values}
\usage{
\method{print}{info_itemgab}(x, out = NULL, len = 5)
}
\description{
Função para exibir os objetos da classe info_itemgab
 
 @param x objeto da classe \code{info_itemgab}
 @param out vetor com o sequencial dos itens que deseja exibir 
    \code{Default: NULL}
 @param len quantidade de itens que deseja exibir. \code{Default: 5}. 
    Para exibir todos os itens pode utilzar \code{len = Inf}.
 @method print info_itemgab
 @S3method print info_itemgab
}
